+++ src/ChooseRequestTypeDialog.js
@@ -10,31 +10,31 @@ import {
 } from '@folio/stripes/components';
 
 import { Loading } from './components';
-import {
-  getRequestTypeOptions,
-} from './utils';
+import { REQUEST_LEVEL_TYPES } from './constants';
+
+export const getRequestTypeError = (requestLevel) => {
+  if (requestLevel === REQUEST_LEVEL_TYPES.TITLE) {
+    return <FormattedMessage id="ui-requests.moveRequest.error.titleLevelRequest" />;
+  }
+
+  return <FormattedMessage id="ui-requests.moveRequest.error.itemLevelRequest" />;
+};
 
 class ChooseRequestTypeDialog extends React.Component {
   static propTypes = {
-    item: PropTypes.object.isRequired,
-    isLoading: PropTypes.bool,
+    isLoading: PropTypes.bool.isRequired,
+    requestLevel: PropTypes.string.isRequired,
     onConfirm: PropTypes.func.isRequired,
     onCancel: PropTypes.func.isRequired,
+    requestTypes: PropTypes.arrayOf(PropTypes.string),
     open: PropTypes.bool,
-  };
-
-  static defaultProps = {
-    isLoading: false,
   };
 
   constructor(props) {
     super(props);
 
-    const { item } = this.props;
-
-    this.options = getRequestTypeOptions(item);
     this.state = {
-      requestType: this.options[0].value,
+      requestType: props.requestTypes[0]?.value,
     };
   }
 
@@ -55,15 +55,19 @@ class ChooseRequestTypeDialog extends React.Component {
       open,
       onCancel,
       isLoading,
+      requestTypes,
+      requestLevel,
     } = this.props;
-    const { requestType } = this.state;
+    const isRequestTypeDisabled = requestTypes.length === 0;
+    const isConfirmButtonDisabled = isRequestTypeDisabled || isLoading;
+    const requestTypesError = isRequestTypeDisabled ? getRequestTypeError(requestLevel) : null;
     const footer = (
       <ModalFooter>
         <Button
           data-test-confirm-request-type
           buttonStyle="primary"
           onClick={this.onConfirm}
-          disabled={isLoading}
+          disabled={isConfirmButtonDisabled}
         >
           <FormattedMessage id="ui-requests.moveRequest.confirm" />
         </Button>
@@ -85,27 +89,28 @@ class ChooseRequestTypeDialog extends React.Component {
         onClose={onCancel}
         footer={footer}
       >
-        {isLoading
-          ? <Loading data-testid="loading" />
-          : this.options.length > 1 ?
-            <Select
-              label={<FormattedMessage id="ui-requests.moveRequest.chooseRequestMessage" />}
-              name="requestType"
-              onChange={this.selectRequestType}
-            >
-              {this.options.map(({ id, value }) => (
-                <FormattedMessage id={id} key={id}>
-                  {translatedLabel => (
-                    <option
-                      value={value}
-                    >
-                      {translatedLabel}
-                    </option>
-                  )}
-                </FormattedMessage>
-              ))}
-            </Select>
-            : <FormattedMessage id="ui-requests.moveRequest.requestTypeChangeMessage" values={{ requestType }} />}
+        {isLoading ?
+          <Loading data-testid="loading" /> :
+          <Select
+            label={<FormattedMessage id="ui-requests.moveRequest.chooseRequestMessage" />}
+            name="requestType"
+            onChange={this.selectRequestType}
+            disabled={isRequestTypeDisabled}
+            error={requestTypesError}
+          >
+            {requestTypes.map(({ id, value }) => (
+              <FormattedMessage id={id} key={id}>
+                {translatedLabel => (
+                  <option
+                    value={value}
+                  >
+                    {translatedLabel}
+                  </option>
+                )}
+              </FormattedMessage>
+            ))}
+          </Select>
+        }
       </Modal>
     );
   }