+++ src/MoveRequestManager.js
@@ -6,12 +6,26 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
-import { stripesConnect } from '@folio/stripes/core';
+import {
+  stripesConnect,
+  stripesShape,
+} from '@folio/stripes/core';
+import {
+  getHeaderWithCredentials,
+} from '@folio/stripes/util';
 
 import ItemsDialog from './ItemsDialog';
 import ChooseRequestTypeDialog from './ChooseRequestTypeDialog';
 import ErrorModal from './components/ErrorModal';
-import { requestTypesByItemStatus } from './constants';
+import { REQUEST_LEVEL_TYPES } from './constants';
+import { getRequestTypeOptions } from './utils';
+
+export const getRequestTypeUrl = (request, okapiUrl, id) => {
+  const url = `circulation/requests/allowed-service-points?requester=${request.requesterId}`;
+  const resourceQuery = request.requestLevel === REQUEST_LEVEL_TYPES.ITEM ? `&item=${id}` : `&instance=${id}`;
+
+  return `${okapiUrl}/${url}${resourceQuery}`;
+};
 
 class MoveRequestManager extends React.Component {
   static propTypes = {
@@ -26,6 +40,7 @@ class MoveRequestManager extends React.Component {
         POST: PropTypes.func.isRequired,
       }),
     }).isRequired,
+    stripes: stripesShape.isRequired,
   };
 
   static manifest = {
@@ -44,6 +59,8 @@ class MoveRequestManager extends React.Component {
     this.state = {
       moveRequest: true,
       moveInProgress: false,
+      requestTypes: [],
+      isRequestTypesLoading: false,
     };
 
     this.steps = [
@@ -69,9 +86,12 @@ class MoveRequestManager extends React.Component {
   }
 
   shouldChooseRequestTypeDialogBeShown = () => {
-    const { selectedItem } = this.state;
-    const { request: { requestType } } = this.props;
-    const requestTypes = this.getPossibleRequestTypes(selectedItem);
+    const { requestTypes } = this.state;
+    const {
+      request: {
+        requestType,
+      },
+    } = this.props;
 
     return !includes(requestTypes, requestType);
   }
@@ -132,27 +152,58 @@ class MoveRequestManager extends React.Component {
     });
   }
 
-  getPossibleRequestTypes(item) {
-    const itemStatus = get(item, 'status.name');
-    return requestTypesByItemStatus[itemStatus] || [];
-  }
-
   onItemSelected = (selectedItem) => {
+    const {
+      request,
+      stripes,
+    } = this.props;
+    const httpHeadersOptions = {
+      ...getHeaderWithCredentials({
+        tenant: stripes.okapi.tenant,
+        token: stripes.store.getState().okapi.token,
+      })
+    };
+    const url = getRequestTypeUrl(request, stripes.okapi.url, selectedItem.id);
+
     this.setState({
-      selectedItem,
-    }, () => this.execSteps(0));
+      isRequestTypesLoading: true,
+      requestTypes: [],
+    });
+
+    fetch(url, httpHeadersOptions)
+      .then(res => {
+        if (res.ok) {
+          return res.json();
+        }
+
+        return Promise.reject(res);
+      })
+      .then((res) => {
+        this.setState({
+          requestTypes: Object.keys(res),
+          selectedItem,
+        }, () => this.execSteps(0));
+      })
+      .catch(() => {
+        this.execSteps(0);
+      })
+      .finally(() => {
+        this.setState({
+          isRequestTypesLoading: false,
+        });
+      });
   }
 
   cancelMoveRequest = () => {
     this.setState({
       moveRequest: true,
       chooseRequestType: false,
+      selectedRequestType: '',
     });
   }
 
   closeErrorMessage = () => {
-    const { selectedItem } = this.state;
-    const requestTypes = this.getPossibleRequestTypes(selectedItem);
+    const { requestTypes } = this.state;
     const state = { errorMessage: null };
 
     if (requestTypes && requestTypes.length > 1) {
@@ -171,10 +222,11 @@ class MoveRequestManager extends React.Component {
     } = this.props;
     const {
       chooseRequestType,
-      selectedItem,
       errorMessage,
       moveRequest,
       moveInProgress,
+      requestTypes,
+      isRequestTypesLoading,
     } = this.state;
 
     return (
@@ -192,8 +244,9 @@ class MoveRequestManager extends React.Component {
           <ChooseRequestTypeDialog
             open={chooseRequestType}
             data-test-choose-request-type-modal
-            item={selectedItem}
-            isLoading={moveInProgress}
+            isLoading={moveInProgress || isRequestTypesLoading}
+            requestTypes={getRequestTypeOptions(requestTypes)}
+            requestLevel={request.requestLevel}
             onConfirm={this.confirmChoosingRequestType}
             onCancel={this.cancelMoveRequest}
           /> }