+++ src/MoveRequestManager.test.js
@@ -9,11 +9,17 @@ import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 
 import '../test/jest/__mock__';
 
-import MoveRequestManager from './MoveRequestManager';
+import MoveRequestManager, {
+  getRequestTypeUrl,
+} from './MoveRequestManager';
 import ItemsDialog from './ItemsDialog';
 import ChooseRequestTypeDialog from './ChooseRequestTypeDialog';
 import ErrorModal from './components/ErrorModal';
-import { requestTypesMap } from './constants';
+import {
+  REQUEST_LEVEL_TYPES,
+  requestTypeOptionMap,
+  requestTypesMap,
+} from './constants';
 
 const labelIds = {
   requestNotAllowed: 'ui-requests.requestNotAllowed',
@@ -37,6 +43,7 @@ const basicProps = {
     instance: {
       title: 'instanceTitle',
     },
+    requestLevel: REQUEST_LEVEL_TYPES.ITEM,
   },
   mutator: {
     move: {
@@ -44,6 +51,19 @@ const basicProps = {
     },
     moveRequest: {
       POST: jest.fn(async () => movedRequest),
+    },
+  },
+  stripes: {
+    okapi: {
+      url: 'okapiUrl',
+      tenant: 'okapiTenant',
+    },
+    store: {
+      getState: () => ({
+        okapi: {
+          token: 'okapiToken',
+        },
+      }),
     },
   },
 };
@@ -106,6 +126,14 @@ jest.mock('./components/ErrorModal', () => jest.fn(({
 )));
 
 describe('MoveRequestManager', () => {
+  beforeEach(() => {
+    global.fetch = jest.fn(() => Promise.resolve({
+      json: () => ({
+        [requestTypesMap.HOLD]: ['id'],
+      })
+    }));
+  });
+
   afterEach(() => {
     jest.clearAllMocks();
     jest.restoreAllMocks();
@@ -146,6 +174,13 @@ describe('MoveRequestManager', () => {
 
   describe('Request type dialog', () => {
     beforeEach(() => {
+      global.fetch = jest.fn(() => Promise.resolve({
+        ok: true,
+        json: () => ({
+          [requestTypesMap.HOLD]: ['id'],
+        })
+      }));
+
       render(
         <MoveRequestManager
           {...basicProps}
@@ -164,40 +199,49 @@ describe('MoveRequestManager', () => {
         onConfirm: expect.any(Function),
         onCancel: expect.any(Function),
         isLoading: false,
-        item: {
-          id: 'selectedItemId',
-          status: {
-            name: 'Checked out',
-          },
-        },
+        requestTypes: [
+          {
+            id: requestTypeOptionMap[requestTypesMap.HOLD],
+            value: requestTypesMap.HOLD,
+          }
+        ],
+        requestLevel: basicProps.request.requestLevel,
       };
 
       await waitFor(() => expect(ChooseRequestTypeDialog).toHaveBeenCalledWith(expect.objectContaining(expectedProps), {}));
     });
 
     it('should hide request type dialog after clicking on "Cancel" button', async () => {
-      const cancelButton = screen.getByTestId(testIds.cancelButton);
-      const chooseRequestTypeDialog = screen.getByTestId(testIds.chooseRequestTypeDialog);
+      await waitFor(() => {
+        const cancelButton = screen.getByTestId(testIds.cancelButton);
+        const chooseRequestTypeDialog = screen.getByTestId(testIds.chooseRequestTypeDialog);
 
-      fireEvent.click(cancelButton);
+        fireEvent.click(cancelButton);
 
-      await waitFor(() => expect(chooseRequestTypeDialog).not.toBeInTheDocument());
+        expect(chooseRequestTypeDialog).not.toBeInTheDocument();
+      });
     });
 
-    it('should trigger "moveRequest.POST" after clicking on "Confirm" button', () => {
-      const confirmButton = screen.getByTestId(testIds.confirmButton);
+    it('should trigger "moveRequest.POST" after clicking on "Confirm" button', async () => {
+      await waitFor(() => {
+        const confirmButton = screen.getByTestId(testIds.confirmButton);
 
-      fireEvent.click(confirmButton);
+        fireEvent.click(confirmButton);
 
-      expect(basicProps.mutator.moveRequest.POST).toHaveBeenCalled();
+        expect(basicProps.mutator.moveRequest.POST).toHaveBeenCalled();
+      });
     });
 
     it('should trigger "onMove" with correct argument after clicking on "Confirm" button', async () => {
-      const confirmButton = screen.getByTestId(testIds.confirmButton);
+      await waitFor(async () => {
+        const confirmButton = screen.getByTestId(testIds.confirmButton);
 
-      fireEvent.click(confirmButton);
+        fireEvent.click(confirmButton);
 
-      await waitFor(() => expect(basicProps.onMove).toHaveBeenCalledWith(movedRequest));
+        await waitFor(() => {
+          expect(basicProps.onMove).toHaveBeenCalledWith(movedRequest);
+        });
+      });
     });
   });
 
@@ -214,6 +258,14 @@ describe('MoveRequestManager', () => {
       const json = jest.fn(() => new Promise(res => res(error)));
 
       beforeEach(async () => {
+        global.fetch = jest.fn(() => Promise.resolve({
+          ok: true,
+          json: () => ({
+            [requestTypesMap.HOLD]: ['holdId'],
+            [requestTypesMap.RECALL]: ['recallId'],
+          })
+        }));
+
         const props = {
           ...basicProps,
           mutator: {
@@ -294,7 +346,7 @@ describe('MoveRequestManager', () => {
       const get = jest.fn(() => '');
       const text = jest.fn(() => new Promise(res => res(error)));
 
-      beforeEach(() => {
+      beforeEach(async () => {
         const props = {
           ...basicProps,
           mutator: {
@@ -314,6 +366,14 @@ describe('MoveRequestManager', () => {
           },
         };
 
+        global.fetch = jest.fn(() => Promise.resolve({
+          ok: true,
+          json: () => ({
+            [requestTypesMap.HOLD]: ['holdId'],
+            [requestTypesMap.RECALL]: ['recallId'],
+          })
+        }));
+
         render(
           <MoveRequestManager
             {...props}
@@ -322,11 +382,13 @@ describe('MoveRequestManager', () => {
 
         const rowButton = screen.getByTestId(testIds.rowButton);
 
-        fireEvent.click(rowButton);
+        await waitFor(() => {
+          fireEvent.click(rowButton);
 
-        const confirmButton = screen.getByTestId(testIds.confirmButton);
+          const confirmButton = screen.getByTestId(testIds.confirmButton);
 
-        fireEvent.click(confirmButton);
+          fireEvent.click(confirmButton);
+        });
       });
 
       it('should trigger "ErrorModal" with correct props', async () => {
@@ -337,6 +399,32 @@ describe('MoveRequestManager', () => {
 
         await waitFor(() => expect(ErrorModal).toHaveBeenCalledWith(expect.objectContaining(expectedProps), {}));
       });
+    });
+  });
+
+  describe('getRequestTypeUrl', () => {
+    const okapiUrl = 'okapiUrl';
+    const requesterId = 'requesterId';
+    const resourceId = 'resourceId';
+
+    it('should return item related link', () => {
+      const request = {
+        requestLevel: REQUEST_LEVEL_TYPES.ITEM,
+        requesterId,
+      };
+      const expectedResult = `${okapiUrl}/circulation/requests/allowed-service-points?requester=${requesterId}&item=${resourceId}`;
+
+      expect(getRequestTypeUrl(request, okapiUrl, resourceId)).toBe(expectedResult);
+    });
+
+    it('should return title related link', () => {
+      const request = {
+        requestLevel: REQUEST_LEVEL_TYPES.TITLE,
+        requesterId,
+      };
+      const expectedResult = `${okapiUrl}/circulation/requests/allowed-service-points?requester=${requesterId}&instance=${resourceId}`;
+
+      expect(getRequestTypeUrl(request, okapiUrl, resourceId)).toBe(expectedResult);
     });
   });
 });